+++ pkg/resources/sql/resources.go
@@ -80,6 +80,8 @@ func init() {
 			ResourceType: DatabaseResourceType,
 			ResourceConfig: base.ResourceConfig{
 				ResourceType: "databases",
+				// A 403 here means the instance is gone; see parentInstanceGone.
+				ReadErrorTreatAsMissing: parentInstanceGone,
 				// A database is nested under its instance:
 				// /projects/{p}/instances/{instance}/databases/{name}
 				ParentResource: &base.ParentResourceConfig{
@@ -103,6 +105,8 @@ func init() {
 			ResourceType: UserResourceType,
 			ResourceConfig: base.ResourceConfig{
 				ResourceType: "users",
+				// A 403 here means the instance is gone; see parentInstanceGone.
+				ReadErrorTreatAsMissing: parentInstanceGone,
 				ParentResource: &base.ParentResourceConfig{
 					ParentType:     "instances",
 					PropertyName:   "instance",
@@ -126,6 +130,8 @@ func init() {
 			OperationConfig: SQLSslCertOperations,
 			ResourceConfig: base.ResourceConfig{
 				ResourceType: "sslCerts",
+				// A 403 here means the instance is gone; see parentInstanceGone.
+				ReadErrorTreatAsMissing: parentInstanceGone,
 				ParentResource: &base.ParentResourceConfig{
 					ParentType:     "instances",
 					PropertyName:   "instance",
@@ -154,6 +160,8 @@ func init() {
 			OperationConfig: SQLBackupRunOperations,
 			ResourceConfig: base.ResourceConfig{
 				ResourceType: "backupRuns",
+				// A 403 here means the instance is gone; see parentInstanceGone.
+				ReadErrorTreatAsMissing: parentInstanceGone,
 				ParentResource: &base.ParentResourceConfig{
 					ParentType:     "instances",
 					PropertyName:   "instance",